fixed minor error in to_hex8

+++ test/utils/mips.hpp
@@ -34,7 +34,7 @@ uint16_t mips_instr_to_opcode(const string &s);
 uint16_t mips_r_instr_to_fncode(const string &s);
 
 // Returns the hex string of uint16_t 'x'
-string to_hex8(uint16_t x);
+string to_hex8(uint32_t x);
 
 // Returns a vector words in sentence 's'
 vector<string> string_break(string s);
